# apps/docs/public/openapi-publisher.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LinkedIn Publisher Backend API",
    "version": "1.0.0",
    "description": "Content creation and publishing platform for LinkedIn posts, carousels, and AI-generated content."
  },
  "components": {
    "securitySchemes": {
      "LinkedInCookie": {
        "type": "apiKey",
        "in": "header",
        "name": "x-linkedin-cookie",
        "description": "LinkedIn session cookie (contains li_at, JSESSIONID, etc.)"
      },
      "LinkedInCsrf": {
        "type": "apiKey",
        "in": "header",
        "name": "x-linkedin-csrf",
        "description": "LinkedIn CSRF token (must match JSESSIONID from cookie)"
      }
    },
...
  },
...
  "paths": {
...
    "/api/posts/upload-images": {
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "uploadImagesToLinkedIn",
        "summary": "Upload images to LinkedIn",
        "description": "Uploads image files to LinkedIn's Ambry CDN and returns media URNs for post attachment.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "images"
                ],
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Image files (max 9, 10MB each)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "mediaUrn": {
                      "type": "string",
                      "description": "Comma-separated media URNs",
                      "example": "urn:li:share:7209384729,urn:li:share:7209384730"
                    },
                    "imageCount": {
                      "type": "integer",
                      "example": 2
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No images provided"
          },
          "401": {
            "description": "LinkedIn credentials not connected"
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "operationId": "listPosts",
        "summary": "List all posts",
        "description": "Returns all posts stored in the database, ordered by creation date.",
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "clx1234..."
                      },
                      "text": {
                        "type": "string",
                        "example": "Excited to share my latest project!"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "draft",
                          "published",
                          "scheduled"
                        ],
                        "example": "published"
                      },
                      "linkedinId": {
                        "type": "string",
                        "nullable": true,
                        "example": "7209384729"
                      },
                      "scheduledAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "createPost",
        "summary": "Create a post",
        "description": "Creates a new post. If status is 'published', immediately publishes to LinkedIn via the Gateway.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Post content text",
                    "example": "Excited to share my latest project!"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published",
                      "scheduled"
                    ],
                    "description": "Post status, default is draft"
                  },
                  "scheduledAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "ISO timestamp for scheduled publishing"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "text",
                      "image",
                      "document"
                    ],
                    "description": "Post type (auto-detected from media)"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Image or PDF files (max 9)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "linkedinId": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Text is required"
          },
          "401": {
            "description": "LinkedIn credentials not connected"
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "operationId": "getPost",
        "summary": "Get a single post",
        "description": "Returns a single post by its database ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Post database ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved successfully"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "operationId": "updatePost",
        "summary": "Update a post",
        "description": "Updates a post's content or status. If transitioning to 'published', sends to LinkedIn.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Post database ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published",
                      "scheduled"
                    ]
                  },
                  "scheduledAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully"
          },
          "404": {
            "description": "Post not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "operationId": "deletePost",
        "summary": "Delete a post",
        "description": "Deletes a post from the database and from LinkedIn if it was published.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Post database ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
    "/api/posts/{id}/publish": {
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "publishPost",
        "summary": "Publish a draft post immediately",
        "description": "Publishes a draft post to LinkedIn immediately, regardless of its current status.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Post database ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mediaUrn": {
                    "type": "string",
                    "description": "Pre-uploaded media URN to attach"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post published successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "example": "published"
                    },
                    "linkedinId": {
                      "type": "string"
                    },
                    "publishedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to publish"
          },
          "404": {
            "description": "Post not found"
          }
        }
      }
    },
...
  }
}